Moved UnixGroup into group.rs and unified loading
