+++ server/lib/src/idm/unix.rs
@@ -1,11 +1,9 @@
-use std::iter;
-
-use kanidm_proto::v1::UnixGroupToken;
 use time::OffsetDateTime;
 use uuid::Uuid;
 
 use kanidm_lib_crypto::CryptoPolicy;
 
+use crate::idm::group::UnixGroup;
 use crate::credential::Credential;
 use crate::modify::{ModifyInvalid, ModifyList};
 use crate::prelude::*;
@@ -192,224 +190,5 @@ impl UnixUserAccount {
             inputs.push(s);
         }
         inputs
-    }
-}
-
-// Would be nice to move this into group.rs to keep the "group" connection.
-// UnixUserAccount can probably be deleted due to the consolidation in Account.
-#[derive(Debug, Clone)]
-pub(crate) struct UnixGroup {
-    pub name: String,
-    pub spn: String,
-    pub gidnumber: u32,
-    pub uuid: Uuid,
-}
-
-macro_rules! try_from_group_e {
-    ($value:expr) => {{
-        // We could be looking at a user for their UPG, OR a true group.
-
-        if !(($value.attribute_equality(Attribute::Class, &EntryClass::Account.to_partialvalue())
-            && $value.attribute_equality(
-                Attribute::Class,
-                &EntryClass::PosixAccount.to_partialvalue(),
-            ))
-            || ($value.attribute_equality(Attribute::Class, &EntryClass::Group.to_partialvalue())
-                && $value.attribute_equality(
-                    Attribute::Class,
-                    &EntryClass::PosixGroup.to_partialvalue(),
-                )))
-        {
-            return Err(OperationError::InvalidAccountState(format!(
-                "Missing {}: {} && {} OR {} && {}",
-                Attribute::Class,
-                Attribute::Account,
-                EntryClass::PosixAccount,
-                Attribute::Group,
-                EntryClass::PosixGroup,
-            )));
-        }
-
-        let name = $value
-            .get_ava_single_iname(Attribute::Name)
-            .map(|s| s.to_string())
-            .ok_or_else(|| {
-                OperationError::InvalidAccountState(format!(
-                    "Missing attribute: {}",
-                    Attribute::Name
-                ))
-            })?;
-
-        let spn = $value
-            .get_ava_single_proto_string(Attribute::Spn)
-            .ok_or_else(|| {
-                OperationError::InvalidAccountState(format!(
-                    "Missing attribute: {}",
-                    Attribute::Spn
-                ))
-            })?;
-
-        let uuid = $value.get_uuid();
-
-        let gidnumber = $value
-            .get_ava_single_uint32(Attribute::GidNumber)
-            .ok_or_else(|| {
-                OperationError::InvalidAccountState(format!(
-                    "Missing attribute: {}",
-                    Attribute::GidNumber
-                ))
-            })?;
-
-        Ok(UnixGroup {
-            name,
-            spn,
-            gidnumber,
-            uuid,
-        })
-    }};
-}
-
-macro_rules! try_from_account_group_e {
-    ($value:expr, $qs:expr) => {{
-        // First synthesise the self-group from the account.
-        // We have already checked these, but paranoia is better than
-        // complacency.
-        if !$value.attribute_equality(Attribute::Class, &EntryClass::Account.to_partialvalue()) {
-            return Err(OperationError::InvalidAccountState(format!(
-                "Missing class: {}",
-                EntryClass::Account
-            )));
-        }
-
-        if !$value.attribute_equality(
-            Attribute::Class,
-            &EntryClass::PosixAccount.to_partialvalue(),
-        ) {
-            return Err(OperationError::InvalidAccountState(format!(
-                "Missing class: {}",
-                EntryClass::PosixAccount
-            )));
-        }
-
-        let name = $value
-            .get_ava_single_iname(Attribute::Name)
-            .map(|s| s.to_string())
-            .ok_or_else(|| {
-                OperationError::InvalidAccountState(format!(
-                    "Missing attribute: {}",
-                    Attribute::Name
-                ))
-            })?;
-
-        let spn = $value
-            .get_ava_single_proto_string(Attribute::Spn)
-            .ok_or_else(|| {
-                OperationError::InvalidAccountState(format!(
-                    "Missing attribute: {}",
-                    Attribute::Spn
-                ))
-            })?;
-
-        let uuid = $value.get_uuid();
-
-        let gidnumber = $value
-            .get_ava_single_uint32(Attribute::GidNumber)
-            .ok_or_else(|| {
-                OperationError::InvalidAccountState(format!(
-                    "Missing attribute: {}",
-                    Attribute::GidNumber
-                ))
-            })?;
-
-        // This is the user private group.
-        let upg = UnixGroup {
-            name,
-            spn,
-            gidnumber,
-            uuid,
-        };
-
-        match $value.get_ava_as_refuuid(Attribute::MemberOf) {
-            Some(riter) => {
-                let f = filter!(f_and!([
-                    f_eq(Attribute::Class, EntryClass::PosixGroup.into()),
-                    f_eq(Attribute::Class, EntryClass::Group.into()),
-                    f_or(
-                        riter
-                            .map(|u| f_eq(Attribute::Uuid, PartialValue::Uuid(u)))
-                            .collect()
-                    )
-                ]));
-                let group_entries: Vec<_> = $qs.internal_search(f)?;
-                let groups: Result<Vec<_>, _> = iter::once(Ok(upg))
-                    .chain(
-                        group_entries
-                            .iter()
-                            .map(|e| UnixGroup::try_from_entry(e.as_ref())),
-                    )
-                    .collect();
-                groups
-            }
-            None => {
-                // No memberof, no groups!
-                Ok(vec![upg])
-            }
-        }
-    }};
-}
-
-impl UnixGroup {
-    pub(crate) fn try_from_account_entry_rw(
-        value: &Entry<EntrySealed, EntryCommitted>,
-        qs: &mut QueryServerWriteTransaction,
-    ) -> Result<Vec<Self>, OperationError> {
-        try_from_account_group_e!(value, qs)
-    }
-
-    pub(crate) fn try_from_account_entry_ro(
-        value: &Entry<EntrySealed, EntryCommitted>,
-        qs: &mut QueryServerReadTransaction,
-    ) -> Result<Vec<Self>, OperationError> {
-        try_from_account_group_e!(value, qs)
-    }
-
-    pub(crate) fn try_from_account_entry<'a, TXN>(
-        value: &Entry<EntrySealed, EntryCommitted>,
-        qs: &mut TXN,
-    ) -> Result<Vec<Self>, OperationError>
-    where
-        TXN: QueryServerTransaction<'a>,
-    {
-        try_from_account_group_e!(value, qs)
-    }
-
-    /*
-    pub fn try_from_account_entry_red_ro(
-        value: &Entry<EntryReduced, EntryCommitted>,
-        qs: &mut QueryServerReadTransaction,
-    ) -> Result<Vec<Self>, OperationError> {
-        try_from_account_group_e!(au, value, qs)
-    }
-    */
-
-    pub(crate) fn try_from_entry_reduced(
-        value: &Entry<EntryReduced, EntryCommitted>,
-    ) -> Result<Self, OperationError> {
-        try_from_group_e!(value)
-    }
-
-    pub(crate) fn try_from_entry(
-        value: &Entry<EntrySealed, EntryCommitted>,
-    ) -> Result<Self, OperationError> {
-        try_from_group_e!(value)
-    }
-
-    pub(crate) fn to_unixgrouptoken(&self) -> UnixGroupToken {
-        UnixGroupToken {
-            name: self.name.clone(),
-            spn: self.spn.clone(),
-            uuid: self.uuid,
-            gidnumber: self.gidnumber,
-        }
     }
 }
